Change original WebMediaPlayer signatures from float to double.
This change puts the final signatures containing doubles in place so that
they can be used by Blink code. After this change lands Blink will be updated
to call these methods instead of the temporary xxxFloat() methods.

BUG=227156
TEST=All existing tests still pass. No user visible behavior has changed.

Review URL: https://codereview.chromium.org/13866046

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@194624 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 39bdde3d92762dcc18f7209e98bf06ab1b82c968

+++ webkit/media/android/webmediaplayer_android.cc
@@ -114,7 +114,7 @@ void WebMediaPlayerAndroid::seekFloat(float seconds) {
   seek(seconds);
 }
 
-void WebMediaPlayerAndroid::seek(float seconds) {
+void WebMediaPlayerAndroid::seek(double seconds) {
   pending_seek_ = seconds;
   seeking_ = true;
 
@@ -133,7 +133,7 @@ void WebMediaPlayerAndroid::setEndTimeFloat(float seconds) {
   setEndTime(seconds);
 }
 
-void WebMediaPlayerAndroid::setEndTime(float seconds) {
+void WebMediaPlayerAndroid::setEndTime(double seconds) {
   // Deprecated.
   // TODO(qinmin): Remove this from WebKit::WebMediaPlayer as it is never used.
 }
@@ -142,7 +142,7 @@ void WebMediaPlayerAndroid::setRateFloat(float rate) {
   setRate(rate);
 }
 
-void WebMediaPlayerAndroid::setRate(float rate) {
+void WebMediaPlayerAndroid::setRate(double rate) {
   NOTIMPLEMENTED();
 }
 
@@ -150,7 +150,7 @@ void WebMediaPlayerAndroid::setVolumeFloat(float volume) {
   setVolume(volume);
 }
 
-void WebMediaPlayerAndroid::setVolume(float volume) {
+void WebMediaPlayerAndroid::setVolume(double volume) {
   NOTIMPLEMENTED();
 }
 
@@ -202,15 +202,15 @@ float WebMediaPlayerAndroid::durationFloat() const {
   return duration();
 }
 
-float WebMediaPlayerAndroid::duration() const {
-  return static_cast<float>(duration_.InSecondsF());
+double WebMediaPlayerAndroid::duration() const {
+  return duration_.InSecondsF();
 }
 
 float WebMediaPlayerAndroid::currentTimeFloat() const {
   return currentTime();
 }
 
-float WebMediaPlayerAndroid::currentTime() const {
+double WebMediaPlayerAndroid::currentTime() const {
   // If the player is pending for a seek, return the seek time.
   if (seeking())
     return pending_seek_;
@@ -244,7 +244,7 @@ float WebMediaPlayerAndroid::maxTimeSeekableFloat() const {
   return maxTimeSeekable();
 }
 
-float WebMediaPlayerAndroid::maxTimeSeekable() const {
+double WebMediaPlayerAndroid::maxTimeSeekable() const {
   // TODO(hclam): If this stream is not seekable this should return 0.
   return duration();
 }
@@ -288,7 +288,7 @@ float WebMediaPlayerAndroid::mediaTimeForTimeValueFloat(float timeValue) const {
   return mediaTimeForTimeValue(timeValue);
 }
 
-float WebMediaPlayerAndroid::mediaTimeForTimeValue(float timeValue) const {
+double WebMediaPlayerAndroid::mediaTimeForTimeValue(double timeValue) const {
   return ConvertSecondsToTimestamp(timeValue).InSecondsF();
 }
