+++ webkit/media/webmediaplayer_impl.h
@@ -96,19 +96,19 @@ class WebMediaPlayerImpl
   virtual bool supportsFullscreen() const;
   virtual bool supportsSave() const;
   virtual void seekFloat(float seconds);
-  virtual void seek(float seconds);
+  virtual void seek(double seconds);
   virtual void setEndTimeFloat(float seconds);
-  virtual void setEndTime(float seconds);
+  virtual void setEndTime(double seconds);
   virtual void setRateFloat(float rate);
-  virtual void setRate(float rate);
+  virtual void setRate(double rate);
   virtual void setVolumeFloat(float volume);
-  virtual void setVolume(float volume);
+  virtual void setVolume(double volume);
   virtual void setVisible(bool visible);
   virtual void setPreload(WebKit::WebMediaPlayer::Preload preload);
   virtual bool totalBytesKnown();
   virtual const WebKit::WebTimeRanges& buffered();
   virtual float maxTimeSeekableFloat() const;
-  virtual float maxTimeSeekable() const;
+  virtual double maxTimeSeekable() const;
 
   // Methods for painting.
   virtual void setSize(const WebKit::WebSize& size);
@@ -128,9 +128,9 @@ class WebMediaPlayerImpl
   virtual bool paused() const;
   virtual bool seeking() const;
   virtual float durationFloat() const;
-  virtual float duration() const;
+  virtual double duration() const;
   virtual float currentTimeFloat() const;
-  virtual float currentTime() const;
+  virtual double currentTime() const;
 
   // Get rate of loading the resource.
   virtual int32 dataRate() const;
@@ -149,7 +149,7 @@ class WebMediaPlayerImpl
   virtual WebKit::WebMediaPlayer::MovieLoadType movieLoadType() const;
 
   virtual float mediaTimeForTimeValueFloat(float timeValue) const;
-  virtual float mediaTimeForTimeValue(float timeValue) const;
+  virtual double mediaTimeForTimeValue(double timeValue) const;
 
   virtual unsigned decodedFrameCount() const;
   virtual unsigned droppedFrameCount() const;
@@ -307,13 +307,13 @@ class WebMediaPlayerImpl
   // SetPlaybackRate(0) is being executed.
   bool paused_;
   bool seeking_;
-  float playback_rate_;
+  double playback_rate_;
   base::TimeDelta paused_time_;
 
   // Seek gets pending if another seek is in progress. Only last pending seek
   // will have effect.
   bool pending_seek_;
-  float pending_seek_seconds_;
+  double pending_seek_seconds_;
 
   WebKit::WebMediaPlayerClient* client_;
